+++ jobs/etl_jobs/external/ppg3/core.py
@@ -3,15 +3,17 @@ from pathlib import Path
 from typing import List, Dict, Optional, Any
 from enum import Enum
 from collections import defaultdict
-from duckdb import DuckDBPyConnection
+import duckdb
+import warnings
 
 from treelib import Tree
 import typer
 import openpyxl
-from config import (REGION_HIERARCHY_TABLE,GCP_PROJECT, BIGQUERY_ANALYTICS_DATASET, ENV_SHORT_NAME,)
+from config import (BASE_TEMPLATE,SOURCE_TABLES,STAKEHOLDER_REPORTS,REGION_HIERARCHY_TABLE,GCP_PROJECT, BIGQUERY_ANALYTICS_DATASET, SHEET_DEFINITIONS, REPORTS)
 from google.cloud import bigquery
 
 from utils.file_utils import slugify
+from utils.data_utils import sanitize_date_fields, sanitize_numeric_types
 
 import os
 from copy import copy
@@ -20,282 +22,16 @@ from dateutil.relativedelta import relativedelta
 import pandas as pd
 import logging
 from utils.duckdb_utils import aggregate_kpi_data, query_yearly_kpi, query_monthly_kpi
-
+from utils.data_utils import build_region_hierarchy
 
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.INFO)
 
-def build_region_hierarchy(
-    project_id: str = GCP_PROJECT,
-    dataset: str = BIGQUERY_ANALYTICS_DATASET,
-    table: str = REGION_HIERARCHY_TABLE,
-) -> Dict[str, Dict]:
-    """
-    Build a nested dict with enhanced academy-department mapping from BigQuery.
-
-    Args:
-        project_id: GCP project ID
-        dataset: BigQuery dataset
-        table: BigQuery table name
-
-    Returns:
-        Dict with structure:
-        {
-            region_name: {
-                "departements": [dept1, dept2, ...],
-                "academies": [acad1, acad2, ...],
-                "academy_departments": {
-                    academy_name: [dept1, dept2, ...]
-                }
-            }
-        }
-
-    Raises:
-        HierarchyError: If hierarchy building fails
-    """
-
-    try:
-        client = bigquery.Client(project=project_id)
-
-        query = f"""
-        SELECT
-            region_name,
-            dep_name,
-            academy_name
-        FROM `{project_id}.{dataset}.{table}`
-        WHERE region_name IS NOT NULL
-        ORDER BY region_name, academy_name, dep_name
-        """
-
-        typer.echo(f"🌍 Querying region hierarchy: {project_id}.{dataset}.{table}")
-
-        rows = client.query(query).result()
-        row_count = 0
-        by_region = {}
-
-        for row in rows:
-            row_count += 1
-            region = row.region_name
-            dep = row.dep_name
-            acad = row.academy_name
-
-            if region not in by_region:
-                by_region[region] = {
-                    "departements": set(),
-                    "academies": set(),
-                    "academy_departments": {},
-                }
-
-            bucket = by_region[region]
-
-            # Add department if present
-            if dep:
-                bucket["departements"].add(dep)
-
-            # Add academy if present
-            if acad:
-                bucket["academies"].add(acad)
-
-            # Map departments to academies
-            if acad and dep:
-                if acad not in bucket["academy_departments"]:
-                    bucket["academy_departments"][acad] = set()
-                bucket["academy_departments"][acad].add(dep)
-
-        # Convert sets to sorted lists for consistent output
-        sorted_by_region = {}
-        for region, scope in sorted(by_region.items()):
-            sorted_by_region[region] = {
-                "departements": sorted(scope["departements"]),
-                "academies": sorted(scope["academies"]),
-                "academy_departments": {
-                    academy: sorted(departments)
-                    for academy, departments in scope["academy_departments"].items()
-                },
-            }
-
-        # Log summary statistics
-        total_regions = len(sorted_by_region)
-        total_academies = sum(len(r["academies"]) for r in sorted_by_region.values())
-        total_departments = sum(
-            len(r["departements"]) for r in sorted_by_region.values()
-        )
-
-        typer.echo(
-            f"✅ Loaded hierarchy: {total_regions} regions, {total_academies} academies, {total_departments} departments"
-        )
-        typer.echo(f"   📊 Total records processed: {row_count:,}")
-
-        return sorted_by_region
-    except Exception as e:
-        typer.secho(f"❌ Failed to build region hierarchy: {e}", fg="red")
-        raise
-
-
-######## base configs
-# GCP_PROJECT = os.environ.get("PROJECT_NAME", "passculture-data-prod")
-# ENV_SHORT_NAME = os.environ.get("ENV_SHORT_NAME", "prod")
-# BIGQUERY_ANALYTICS_DATASET = f"analytics_{ENV_SHORT_NAME}"
-
-
-# REGION_HIERARCHY_TABLE = "region_department"
-
-# TAB_NAME_MAX_LENGTH = 31
-
-# ALL_TABLES = {
-#     "individual": {"table_name": f"{table_prefix}individual", "tab": "Part individuelle", "scope": "individual"},
-#     "collective": {"table_name": f"{table_prefix}collective", "tab": "Part collective", "scope": "collective"},
-#     "top_individual_offer": {"table_name": f"{table_prefix}top_offer", "tab": "TOP 50 offres - individuel", "scope":"individual"},
-#     "top_individual_offer_category": {"table_name": f"{table_prefix}top_offer_category", "tab": "TOP 50 offres par catégorie - individuel", "scope":"individual"},
-#     "top_individual_venue": {"table_name": f"{table_prefix}top_venue", "tab": "TOP 50 lieux - individuel", "scope":"individual"},
-# }
-
-
-
-
-STAKEHOLDER_REPORTS = {
-    "ministere": ["national_summary"],
-    "drac": ["regional_summary", "departemental_detail", "academy_detail"]
-}
-
-# BASE_TEMPLATE = Path("./jobs/etl_jobs/external/ppg/templates/PPG_Template_v4.xlsx")
-BASE_TEMPLATE = Path("./templates/export_template_v4.xlsx")
-
-TEMPLATE_DEFAULT = Path("./templates/export_template_v2.xlsx")
-
-
-
-# data sources
-table_prefix = "external_reporting"
-
-SOURCE_TABLES = {
-    "individual": {"dataset": BIGQUERY_ANALYTICS_DATASET, "table": f"{table_prefix}_individual"},
-    "collective": {"dataset": BIGQUERY_ANALYTICS_DATASET, "table": f"{table_prefix}_eac"},
-    "top_offer": {"dataset": BIGQUERY_ANALYTICS_DATASET, "table": f"{table_prefix}_top_offer"},
-    "top_offer_category": {"dataset": BIGQUERY_ANALYTICS_DATASET, "table": f"{table_prefix}_top_offer_category"},
-    "top_venue": {"dataset": BIGQUERY_ANALYTICS_DATASET, "table": f"{table_prefix}_top_venue"},
-    "top_ac": {"dataset": BIGQUERY_ANALYTICS_DATASET, "table": f"{table_prefix}_top_ac"},
-    "top_format": {"dataset": BIGQUERY_ANALYTICS_DATASET, "table": f"{table_prefix}_top_format"},
-}
-class SheetType(Enum):
-    KPIS = "kpis"
-    TOP = "top"
-    LEXIQUE = "lexique"
-
-SHEET_DEFINITIONS = {
-    "individual_kpis": {
-        "type": SheetType.KPIS,
-        "template_tab": "template_individuel",
-        "source_table": "individual",
-    },
-    "collective_kpis": {
-        "type": SheetType.KPIS,
-        "template_tab": "template_collectif",
-        "source_table": "collective",
-    },
-    "lexique": {
-        "type": SheetType.LEXIQUE,
-        "template_tab": "Lexique",
-        "source_table": None,
-    },
-    "top_offer": {
-        "type": SheetType.TOP,
-        "template_tab": "Top 50 offres",
-        "source_table": "top_offer",
-        "top_n": 50,
-        "select_fields": ["partition_month","offer_category_id","offer_subcategory_id","offer_name","total_booking_amount","total_booking_quantity"],
-        "order_by": ["total_booking_amount"]
-    },
-    "top_offer_category": {
-        "type": SheetType.TOP,
-        "template_tab": "Top 50 par catégorie",
-        "source_table": "top_offer_category",
-        "top_n": 50,
-        "select_fields": ["partition_month","offer_category_id","offer_subcategory_id","offer_name","total_booking_amount","total_booking_quantity"],
-        "order_by": ["offer_category_id","total_booking_amount"]
-    },
-    "top_venue": {
-        "type": SheetType.TOP,
-        "template_tab": "Top 50 lieux",
-        "source_table": "top_venue",
-        "top_n": 50,
-        "select_fields": ["partition_month","venue_name","offerer_name"],
-        "order_by": ["total_venue_booking_amount_ranked"]
-    },
-    "top_ac": {
-        "type": SheetType.TOP,
-        "template_tab": "Top 50 acteurs culturels",
-        "source_table": "top_ac",
-        "top_n": 50,
-        "select_fields": ["partition_month","offerer_name","total_number_of_tickets"],
-        "order_by": ["total_booking_amount"]
-    },
-    "top_format": {
-        "type": SheetType.TOP,
-        "template_tab": "Top 5 formats",
-        "source_table": "top_format",
-        "top_n": 5,
-        "select_fields": ["partition_month","collective_offer_format","total_booking_amount","total_number_of_tickets"],
-        "order_by": ["total_booking_amount"]
-    },
-}
-
-
-REPORTS = {
-    "national_summary": {
-        "sheets": [
-            {"definition": "lexique"},
-            {"definition": "individual_kpis", "filters": {"scope": "individual", "scale": "national"}},
-            {"definition": "top_offer", "filters": {"scope": "individual", "scale": "national"}},
-            {"definition": "top_offer_category", "filters": {"scope": "individual", "scale": "national"}},
-            {"definition": "top_venue", "filters": {"scope": "individual", "scale": "national"}},
-            {"definition": "collective_kpis", "filters": {"scope": "collective", "scale": "national"}},
-            {"definition": "top_ac", "filters": {"scope": "collective", "scale": "national"}},
-            {"definition": "top_format", "filters": {"scope": "collective", "scale": "national"}},
-        ]
-    },
-    "region_summary": {
-        "sheets": [
-            {"definition": "lexique"},
-            {"definition": "individual_kpis", "filters": {"scope": "individual", "scale": "national"}},
-            {"definition": "individual_kpis", "filters": {"scope": "individual", "scale": "region"}},
-            {"definition": "top_offer", "filters": {"scope": "individual", "scale": "region"}},
-            {"definition": "top_offer_category", "filters": {"scope": "individual", "scale": "region"}},
-            {"definition": "top_venue", "filters": {"scope": "individual", "scale": "region"}},
-            {"definition": "collective_kpis", "filters": {"scope": "collective", "scale": "national"}},
-            {"definition": "collective_kpis", "filters": {"scope": "collective", "scale": "region"}},
-            {"definition": "top_ac", "filters": {"scope": "collective", "scale": "region"}},
-            {"definition": "top_format", "filters": {"scope": "collective", "scale": "region"}},
-        ]
-    },
-    "academy_detail": {
-        "sheets": [
-            {"definition": "lexique"},
-            {"definition": "collective_kpis", "filters": {"scope": "collective", "scale": "region"}},
-            {"definition": "collective_kpis", "filters": {"scope": "collective", "scale": "academie"}},
-            {"definition": "top_ac", "filters": {"scope": "collective", "scale": "academie"}},
-            {"definition": "top_format", "filters": {"scope": "collective", "scale": "academie"}},
-        ]
-    },
-    "department_detail": {
-        "sheets": [
-            {"definition": "lexique"},
-            {"definition": "individual_kpis", "filters": {"scope": "individual", "scale": "region"}},
-            {"definition": "individual_kpis", "filters": {"scope": "individual", "scale": "departement"}},
-            {"definition": "top_offer", "filters": {"scope": "individual", "scale": "departement"}},
-            {"definition": "top_offer_category", "filters": {"scope": "individual", "scale": "departement"}},
-            {"definition": "top_venue", "filters": {"scope": "individual", "scale": "departement"}},
-        ]
-    },
-}
-
-
 
 class StakeholderType(Enum):
     MINISTERE = "ministere"
     DRAC = "drac"
-
 
 @dataclass
 class Stakeholder:
@@ -391,18 +127,22 @@ class Sheet:
         except Exception as e:
             logger.warning(f"Failed to resolve dimension context: {e}")
             return None
+        
+    def __post_init__(self):
+        if not self.context:
+            self.context = self.get_dimension_context()
 
-    def preprocess(self, ds: str):
-        """Placeholder for preprocessing - now handled by ReportOrchestrationService."""
-        pass
+    # def preprocess(self, ds: str):
+    #     """Placeholder for preprocessing - now handled by ReportOrchestrationService."""
+    #     pass
         
-    def fill_data(self, duckdb_conn: DuckDBPyConnection, ds: str):
-        """Placeholder for data filling - now handled by ReportOrchestrationService.""" 
-        pass
+    # def fill_data(self, duckdb_conn: duckdb.DuckDBPyConnection, ds: str):
+    #     """Placeholder for data filling - now handled by ReportOrchestrationService.""" 
+    #     pass
         
-    def postprocess(self):
-        """Placeholder for postprocessing."""
-        pass
+    # def postprocess(self):
+    #     """Placeholder for postprocessing."""
+    #     pass
 
 
 
@@ -498,7 +238,7 @@ class Report:
             self.workbook._sheets.remove(ws)
             self.workbook._sheets.insert(idx, ws)
     
-    def build(self, ds: str, duckdb_conn: DuckDBPyConnection):
+    def build(self, ds: str, duckdb_conn: duckdb.DuckDBPyConnection):
         """
         Build the report workbook using the new service architecture.
         
@@ -604,3 +344,166 @@ class ReportPlanner:
                     )
 
         return jobs
+    
+
+class ExportSession:
+    """Manages the entire export session with proper resource cleanup."""
+    def __init__(self, ds: str):
+        self.ds = ds
+        self.tables_to_load = SOURCE_TABLES
+        self.conn: Optional[duckdb.duckdb.DuckDBPyConnection] = None
+
+    def __enter__(self):
+        return self
+        
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        if self.conn:
+            self.conn.close()
+
+    
+    def load_data(self):
+        """Load tables based on self.scope."""
+        
+        client = bigquery.Client(project=GCP_PROJECT)
+        self.conn = duckdb.connect(":memory:")
+        
+        for _, config in self.tables_to_load.items():
+            self._load_table(client,config)
+    
+    def _load_table(self, client: bigquery.Client, config: Dict):
+        """Load a single table into DuckDB."""
+        table_name = config['table']
+        dataset = config.get('dataset', BIGQUERY_ANALYTICS_DATASET)
+        project_id = config.get('project', GCP_PROJECT)
+        typer.echo(f"Loading {table_name} from BigQuery...")
+        query = f"SELECT * FROM `{project_id}.{dataset}.{table_name}`"
+        with warnings.catch_warnings():
+            warnings.filterwarnings("ignore", message=".*BigQuery Storage module not found.*")
+            df = client.query(query).to_dataframe()
+            typer.echo(f"{table_name} dtypes: {df.dtypes}")
+        if df.empty:
+            typer.echo(f"Warning: {table_name} returned no data")
+            return
+        
+        df = sanitize_date_fields(df, "partition_month")
+        df = sanitize_numeric_types(df)
+            
+        typer.echo(f"Loaded {len(df):,} records for {table_name}")
+        
+        # Register and create table
+        self.conn.register(f"{table_name}_tmp", df)
+        self.conn.execute(f"CREATE TABLE {table_name} AS SELECT * FROM {table_name}_tmp")
+        
+        # Create indexes based on available columns
+        self._create_indexes(table_name, df.columns.tolist())
+        typer.echo(f"Created {table_name} table with indexes")
+
+    
+    def _create_indexes(self, table_name: str, columns: List[str]):
+        """Create indexes on commonly queried columns with error recovery."""
+        indexes_created = []
+        indexes_failed = []
+        
+        typer.secho(f"🔍 Creating indexes for table: {table_name}", fg="cyan")
+        
+        # List of indexes to attempt
+        index_operations = []
+        
+        # Core filtering columns
+        core_filter_columns = ['dimension_name', 'dimension_value', 'partition_month']
+        kpi_columns = ['kpi_name']
+        
+        if all(col in columns for col in core_filter_columns):
+            index_operations.append({
+                "name": "core_filters",
+                "sql": f"CREATE INDEX IF NOT EXISTS idx_{table_name}_core_filters ON {table_name} (dimension_name, dimension_value, partition_month)",
+                "description": "Core filters composite index"
+            })
+        
+        if all(col in columns for col in core_filter_columns + kpi_columns):
+            index_operations.append({
+                "name": "kpi_filters", 
+                "sql": f"CREATE INDEX IF NOT EXISTS idx_{table_name}_kpi_filters ON {table_name} (kpi_name, dimension_name, dimension_value, partition_month)",
+                "description": "KPI-specific composite index"
+            })
+        
+        # Individual indexes
+        for col in ['partition_month', 'dimension_name','dimension_value', 'kpi_name']:
+            if col in columns:
+                index_operations.append({
+                    "name": col,
+                    "sql": f"CREATE INDEX IF NOT EXISTS idx_{table_name}_{col} ON {table_name} ({col})",
+                    "description": f"Individual index on {col}"
+                })
+        
+
+        # Execute each index creation with individual error handling
+        for idx_op in index_operations:
+            try:
+                self.conn.execute(idx_op["sql"])
+                indexes_created.append(idx_op["name"])
+                typer.echo(f"   ✓ {idx_op['description']}")
+            except Exception as e:
+                indexes_failed.append(idx_op["name"])
+                if "INTERNAL Error" in str(e):
+                    typer.secho(f"   ⚠ {idx_op['description']} - skipped (DuckDB internal error)", fg="yellow")
+                else:
+                    typer.secho(f"   ❌ {idx_op['description']} - failed: {str(e)[:50]}...", fg="red")
+        
+        # Summary
+        if indexes_created:
+            typer.secho(f"📊 {table_name}: Created {len(indexes_created)} indexes successfully", fg="green")
+        if indexes_failed:
+            typer.secho(f"⚠ {table_name}: {len(indexes_failed)} indexes failed (table will still work)", fg="yellow")
+        
+
+    def process_stakeholder(self, stakeholder_type: str, name: str, output_path: Path, ds: str):
+        """Process reports for a given stakeholder using new service architecture."""
+        stakeholder = Stakeholder(
+            type=StakeholderType.MINISTERE if stakeholder_type == "ministere" else StakeholderType.DRAC,
+            name="Ministère" if stakeholder_type == "ministere" else name,
+        )
+        typer.secho(f"➡️ Processing reports for {stakeholder.type.value} - {stakeholder.name}", fg="cyan")
+        
+        # Generate reports based on stakeholder's desired reports
+        planner = ReportPlanner(stakeholder)
+        report_jobs = planner.plan_reports()
+        
+        successful_reports = 0
+        total_reports = len(report_jobs)
+        
+        typer.secho(f"📋 Planning to generate {total_reports} reports", fg="cyan")
+        
+        for job in report_jobs:
+            try:
+                # Create Report instance
+                report = Report(
+                    report_type=job["report_type"],
+                    stakeholder=stakeholder,
+                    base_template_path=BASE_TEMPLATE,
+                    output_path=output_path / job["output_path"],
+                    context=job["context"]
+                )
+                
+                # Build and save report using new architecture
+                report.build(ds, self.conn)
+                report.save()
+                
+                successful_reports += 1
+                
+            except Exception as e:
+                typer.secho(f"❌ Failed to generate report {job['output_path']}: {e}", fg="red")
+                logger.error(f"Report generation failed for {job['output_path']}: {e}")
+                # Continue with next report instead of failing completely
+        
+        # Final summary
+        if successful_reports == total_reports:
+            typer.secho(f"✅ All {total_reports} reports generated successfully for {stakeholder.name}", fg="green")
+        elif successful_reports > 0:
+            typer.secho(
+                f"⚠️ {successful_reports}/{total_reports} reports generated successfully for {stakeholder.name}", 
+                fg="yellow"
+            )
+        else:
+            typer.secho(f"❌ No reports generated successfully for {stakeholder.name}", fg="red")
+
